Internal API addition for IE driver to get browser window handle
This commit introduces an API for getting the browser window handle. In
IE6, this would be the top-level window handle. In IE7 and above, this
will be the handle of the tab window. This is separate from the window
which renders the content, which has a window class of
'InternetExplorer_Server'. There is a need to locate both window handles
distinctly in the internal operation of the IE driver.