Disable data reporting features in default Firefox profile used by FirefoxDriver.
This prevents "Firefox automatically sends some data to Mozilla so that we can improve your experience" [Choose What I Share] prompt from showing up at more or less random times during the session - which started happening (sooner in the session?) with Firefox version 34.0.
When the prompt shows up it 1) steals focus - which breaks typing and possibly some other WebDriver commands and 2) reduces size of the viewport/screenshot.

Note that setting either datareporting.policy.dataSubmissionPolicyAccepted:true or datareporting.policy.dataSubmissionPolicyBypassAcceptance:true (or, possibly, setting datareporting.policy.dataSubmissionPolicyResponseType to "accepted-info-bar-dismissed" or "accepted-info-bar-button-pressed") would be a more robust way to get rid of the prompt as it would indicate that the user or (in case of dataSubmissionPolicyBypassAcceptance) someone on behalf of the user already accepted the policy and there's no need to ask again. We do not enable either of those settings here by default as FirefoxDriver users deserve to be notified of anything that their browsers would send to Mozilla.
As or Firefox 34 the datareporting.policy covers Firefox Health Reporter (FHR), Telemetry, and Crash reporter (crash reporter settings can be found in application.ini, webapprt.ini). But we don't want to blindly accept any new kinds of data Firefox may possibly want to send to Mozilla in the future.
Disabling all 3 of FHR, Telemetry and Crash reporter *should be* enough to prevent the prompt from appearing, at least as of Firefox 34.
https://bugzilla.mozilla.org/show_bug.cgi?id=873054 and https://bugzilla.mozilla.org/show_bug.cgi?id=862563 have plenty of related information.