Fixed imports for str, StringIO, basestring
Used six package to make the code more pythonic.  It’s cleaner and
everything behaves as expected.  With the prior method, base types were
being overridden.  This is a bad idea because suddenly a person reading
the code has to read the entire code just to figure out what something
is actually doing rather than just knowing what the default behavior
is.