webdriverjs mocha: wait for 'done' callbacks to be invoked
Old code was passing the promise reject method as the test done callback.
The promise representing the 'it' was always completed immediately.  This
change makes it so if a test function is expecting a done callback, the
callback is given control over resolving the promise.

Add tests for checking the "done" callback on an asynchronous 'it'
invocation causes the system to wait correctly when using the webdriverjs
wrappers for the mocha 'it'; and tests for the interaction of 'done' and
control flow.

Several tests were passing the httpserver "start" method as a "before"
callback.  Since this method takes an optional port number, it has a
formal parameter, and so the updated 'it' infrastructure assumed these
functions were async and expected a "done" callback to be passed in.  That
didn't work.

Plus a smattering of documentation cleanups around the JavaScript code.

Signed-off-by: Jason Leyba <jmleyba@gmail.com>