webdriverjs mocha: handle 'done' callbacks correctly
Old code was passing the promise reject method as the done callback.  The
promise representing the 'it' was always completed immediately.  This
change makes it so if a callback is expecting a done callback, the
callback is given control over resolving the promise.  (Otherwise the
existing code which completes the promise immediately or defers to a
returned promise is used.)

Cleanup the code a bit (move a nested function out one level, rename the
"done" promise that shadowed containing functions "done" callback, add
some comments, give some of the anonymous functions names so that
stacktraces are easier to read).