Fix tests to work with async done callback support
Several tests were passing the httpserver "start" method as a "before"
callback.  Since this method takes an optional port number, it has a
formal parameter, and so the updated 'it' infrastructure assumed these
functions expected a "done" callback to be passed in.  That didn't work.

Also fix a small grammatical error in a new comment I had added