Remove automatic installation of the SafariDriver in Java.
Moving forward, users will be expected to manually install the
extension. This change is response to architectural changes introduced
with Safari 7 that broke auto-installation.

Starting with the next Selenium release, SafariDriver.safariextz will be
posted along with the other Selenium release artifacts. The extension
does not currently auto-update; we may yet add support for this.

This change also removes support for installing additional extensions
via SafariOptions (for the same reasons we no longer support installing
the driver itself). All extension related methods in SafariOptions have
been deprecated, are no-ops, and will be removed in the next+1 release.

Fixes issue 8514