dotnet: make PhantomJS driver more cross-platform aware
On Linux the filesystem is case-sensitive and executables
don't end with ".exe". This change modifies the PhantomJSDriver
to search for "phantomjs" on Unix systems while keeping the
filename "PhantomJS.exe" on Windows. This allows to use the
default constructor of the PhantomJSDriver class on Linux.
Previously using the default c'tor resulted in an error that
"PhantomJS.exe" can't be found.

Signed-off-by: Andreas Tolfsen <ato@mozilla.com>