.NET PageFactory refactor
This commit is a fairly large refactor of the .NET PageFactory
implementation. It includes a number of changes to make it more extensible
and more robust.

First, instead of the hand-rolled proxy classes that had been
used previously, this commit introduces proxy objects based on
System.Runtime.Remoting.Proxies.RealProxy, which allows for more
flexibility in the types of objects proxied.

Second, this commit now allows users to provide their own
implementation engines for examining attributes of Page Object classes
and determining what course to take. To create a new implementation
the user must provide an implementation of the IPageObjectMemberDecorator
interface, which determines which attributes to examine fields and
properties for, and how to create the proxy objects assigned to the
members.

Third, this commit introduces a new IElementLocator interface to Page
Object implementations, and deprecates the IElementLocatorFactory
interface. The IElementLocator interface is substantially similar to the
now-deprecated IElementLocatorFactory, but requires implementations to
provide a property to retrieve the ISearchContext implementation used to
actually find elements. The implications of this change are that
IElementLocator implementations will need to keep a reference to the
IWebDriver or IWebElement they are using. Recommended practice is for this
context to be passed in via a the IElementLocator implementation's
constructor, and this technique is demonstrated in the source code to the
two implementations provided in the regular .NET language bindings
releases.

Finally, this commit also introduces a breaking change in
IElementLocatorFactory. Developers of classes that implement this
interface will now be required to implement the CreateLocator method,
which will return a valid IElementLocator implmentation. This is to
prepare for the removal of the IElementLocatorFactory interface and its
implmentations in a future release.

Signed-off-by: Jim Evans <james.h.evans.jr@gmail.com>