Adding FindsByAllAttribute to .NET PageFactory implementation
This allows the user to specify that an element must match the criteria of
all of the FindsBy attributes a property or field is marked with in order
to be found by the PageFactory. Previously, a class member marked with
multiple FindsBy attributes would find an element matching any of the
attributes' criteria. Because of the nature of this change, you cannot use
FindsBySequence and FindsByAll on the same class member. Attempting to do
so will result in an exception when InitElements is called.