Disable https://wiki.mozilla.org/Advocacy/heartbeat Firefox 37+ feature in browser sessions started by FirefoxDriver.
According to the documentation,
"Every day a random subset of users are offered a rating widget",
which doesn't sound great for automated tests. We have not really
recorded any harmful effects (such as flakiness) of this feature
on tests, but I thought this change should keep everyone safe from that.

This is the only way to disable the feature that we found so far.