Generalize SocketLock to account for Chrome port race condition
When the Chrome browser boots up via the Ruby library, it attempts
to find a free port and launch the bridge on that port. However, it's
possible that, if you're running tests in parallel, two bridges will
try to bind to the same port at the same time. This will cause a very
obtuse "end of file" error as documented [here][1].

This patch is based on one initially proposed in the original
[bug report][2]. However, it goes a step further, taking the technical
comments into account, and generalizing the `SocketLock` class used
in the Firefox launcher to assist.

[1]: https://github.com/grosser/parallel_tests/issues/322
[2]: https://code.google.com/p/selenium/issues/detail?id=8241

Signed-off-by: Alexei Barantsev <barancev@gmail.com>