Adding ability to get wrapped IWebElement from SelectElement in .NET
When using a SelectElement wrapper class to manipulate <select> elements,
the user previously had to keep a separate reference to the IWebElement
representing the actual element. With this commit, the SelectElement class
now implements IWrapsElement, exposing the WrappedElement property, which
gives the user the ability to access the underlying IWebElement. This
fixes issue #350.