Reset browser window handle on failure to attach via Active Accessibility
When launching a new instance, if the driver attempts to use Active
Accessibility to attach to the instance, and the attach fails, the
driver should attempt to attach using the ShellWindows API. However,
this fallback is skipped if the window has already been found. Resetting
the found window handle to NULL allows the driver to proceed with the
second attempt to attach.