In SafariDriver's page script, copy window properties to goog.global.
When the page script is injected, it's called with an empty 'this'
object to avoid polluting the global namespace of the page under
test. However, some Closure libraries look for properties on
goog.global and fall back to worse or slower versions if those
properties are absent.

The specific motivating case is that if goog.async.nextTick can't find
goog.global.MessageChannel, it will fall back to a method that injects
an invisible iframe into the DOM. The page script's use of nextTick
would cause it to create an iframe in the DOM of the page under test,
which we'd like to avoid.