Adding Python code to launch BlackBerry browser automatically
The BlackBerry10 browser doesn't launch automatically with WebDriver, which
makes automated testing problematic. This patch enables the automatic
launching of the BlackBerry10 browser, if the user has the proper env.
setup.

This patch assumes the IP/port of the device. The user must provide location
of required tools and device password.

Change-Id: Id7d5845242673fb94fe617965ee1ffb4828489ae