Creating new Visual Studio solution files
This separates out the various C++ projects and .NET projects into logical
solution files based on the purpose of the projects. There is now one
solution for the (now deprecated) Firefox native events libraries; one for
the C++ IE driver code; and one for the .NET langauge bindings, written in
C#. This will allow people to more easily isolate the changes they're
working on, and allow for an easier upgrade path to future versions of
Visual Studio.