Fixing style issues + making BB tools easier
Fixed up python styling issues (pep8). Added a better description of the args.

Realized that blackberry tools when installed with the WebWorks SDK are in the path
so changed it so those are handled by default. Also made the 'typical' ip address
used to be the default, now the user only needs to supply a device password if
everything is setup in the standard way.

Added some directions/hints on getting the webworks SDK, location should never
change.

Change-Id: Ia558b0777b0abbcf7567f73943cf5ec83403064d