set window-focus before clicking on item
We are using Java-Selenium tests with Firefox driver for testing a
web-application based on Vaadin. The Firefox driver correctly filled some
Vaadin textboxes with the desired text, however the text was lost when it
was saved by using an appropriate save-button in the web-application.
This problem in the tests only occured when the Selenium tests were
debugged in Eclipse (e.g. step-by-step), in the Release version of the
tests, everything worked fine.

The problem is, that every click on the debugger (in our case
Eclipse) caused that Firefox lost Window-Focus (Windows sent
WM_ACTIVATE-WA_INACTIVE to Firefox process) which itself caused Vaadin-
Javascripts to disable reading the content of the textboxes.

(By using mouse and keyboard it is impossible to cause such a state
manually)

This Firefox Driver patch sends a "WM_ACTIVATE-WA_CLICKACTIVE" and
"WM_SETFOCUS" message to Firefox before clicking/double-clicking on
an item.