Allow IE driver to work with transparent file upload elements.
Many modern JavaScript widget frameworks hide <input type="file"> elements
so as to make their UIs "prettier." The IE driver would see these elements
as not visible, and thus not interactible. This commit allows users to use
sendKeys to upload files when this type of file upload element exists,
bringing the driver in line with what is allowed by the Firefox driver.