Minor DefaultFieldDecorator improvement
Visibility of the "isDecoratableList" method has been changed to
Protected.

There are lots of Selenium-based projects that use PageFactory and customized decorating engines that
are similar as engines of DefaultFieldDecorator. It means that this class can be extended and its
methods can be reused or overridden. Actually almost all methods of DefaultFieldDecorator are public
or protected except "isDecoratableList". So why it can't be protected too?