Don't convert non String types to lower in python
It appears that the original issue in
https://code.google.com/p/selenium/issues/detail?id=5955 has returned.
If webdriver returns integer and boolean values as integera and boolean
types, calling .lower() is not valid.

This commit checks the type first, before checking if it's a boolean
being returned as a string (so the fix is slightly different than in the
issue linked above). It's also a good piece of defensive coding,
regardless.