Add test annotation @NotYetImplemented
This feature is meant to be used with test classes with
@RunWith(SeleniumTestRunner.class)
It makes the runner executes the test method, and if
the method fails (as expected) then the build will be green,
and if the method succeeds, then the build is red complaining
"Test is marked as @NotYetImplemented but already works".

This feature ensures that anything marked with @NotYetImplemented
is still not implemented, and it is helpful to make
the test suite annotations correct.

This commit applies it to tests in TypingTest running with HtmlUnit
(which were previously @Ignore), but it can be applied to other
classes or drivers.

Signed-off-by: Luke Inman-Semerau <luke.semerau@gmail.com>