Adding ability to set alert credentials via remote
This change is twofold. First, it adds a method to the Alert interface in
the Java language bindings to enable the ability to set credentials
without implicitly clicking OK on the authentication alert. The first
change enables the second, which is to expose the 'set alert credentials'
wire protocol end point to users of the Java remote server.