Ignoring failing cookie test for IE
The IE driver can correctly handle cookies with the 'secure' flag set,
but the browser throws up a mandatory, unignorable interstitial page when
the site in question uses an untrusted SSL certificate. Until and unless
a technique for dealing with this interstitial page automatically can be
found, this test, or any test requiring browsing via https, must be
ignored in this test suite.