Enabling fast failure for cookie manipulation in IE
The refactor of cookie handling for the IE driver introduces an
incompatibility with the 64-bit IE driver and IE 10 and 11 running on
64-bit Windows. As is the case with sending keystrokes and creating
screenshots, a Windows hook procedure is now used for getting and setting
cookies in IE. That means that in IE 10 and 11 on 64-bit Windows, where
the content rendering process is still 32-bit, you **must** use the 32-bit
IEDriverServer.exe in order to manipulate cookies.

This commit will now cause exceptions to be thrown if you attempt to set
or get cookies using the 64-bit driver against a 32-bit version of IE (or
vice versa), but in particular, this will affect users who mistakenly try
to use the 64-bit executable on IE 10 or 11 in 64-bit Windows.