Adding logging messages for IE driver for bitness mismatches
This commit adds logging messages at the warning level whenever a mismatch
in bitness is detected between the browser and driver. This is a
particularly bad problem in cases where users have 64-bit Windows, and
(incorrectly) assume they automatically want to use the 64-bit
IEDriverServer.exe when running against IE 10 or 11. Since the process
that actually renders content in IE 10 and 11 is almost always 32-bit,
even on 64-bit Windows, it is almost always the proper decision to use
the 32-bit IE driver. It is now very clear from log messages when this
state of affairs exists.