Internal changes to JSON wire protocol handling in .NET bindings
This commit does away with extraneous JsonConverter classes used with the
Json.NET serialization library. The changes should be merely internal, but
companion projects that incorrectly rely on implementation details of the
.NET bindings may find these changes to be breaking. Please adjust
accordingly. This prepares the way for future changes to enable the .NET
bindings to correctly communicate using the W3C dialect of the JSON Wire
Protocol.