Add FluentWait.withMessage with string supplier
Add a withMethod(Supplier<String> messgaeSupplier) to be able to evaluate the message when the wait is over, thus enabling developers to show more info on the context of failure through closures. Internally, the FluentWait class is using a supplier.
withMessage(String message) method kept for backward compatibility creates a supplier that return the given string.