Updating screenshot code to prevent resizing if window is large enough.
In the previous screenshot update, the driver would incorrectly resize
the IE window smaller if the document was smaller in height or width
than the current window size. This could cause redraws of the page
being viewed such that after resizing, the screenshot still did not
display the entire page. This commit resolves that issue, and also takes
into account the size of the scroll bars when creating the screenshot
image. NOTE: There is a *very* slight chance that resizing the window
so there are no longer scroll bars to be displayed *might* still cause
layout redraws such that the screenshot does not show the entire DOM
after the resize. Since we should now always be expanding the window size,
never contracting it, this is a corner case that explicitly will *not* be
fixed. Any issue reports describing this corner case will be closed
without action. Fixes issue #1085.