Updating screenshot code to retain scroll bars in required directions.
When an IE window needs to be resized to obtain the full-DOM screenshot,
if it has scroll bars in either direction, those scroll bars need to be
present in the resized window, so that layout recalculation doesn't occur.
In practice, this means reducing the window size by 2 pixels in the
appropriate direction before taking the screenshot. The resulting scroll
bars are cropped from the image.