Add HTML5 functionality to .NET bindings
This includes being able to access appCache status, local and session
storage, and geolocation. Each feature has a role-based marker interface
(IHasApplicationCache, IHasWebStorage, or IHasLocationContext), and the
implementation has a boolean property to query if the driver supports that
feature. Attempting to use the feature for a driver which does not support
it will throw an exception.

Signed-off-by: Jim Evans <james.h.evans.jr@gmail.com>