Updating .NET Firefox driver to use standard init and terminate
This commit modifies the Marionette driver to test for initialization
using a URL poll instead of a hard-coded sleep. We poll for the URL to
delete a nonexistent session. This still throws an error, but one that
still lets us know that the HTTP server is available and responding to
HTTP requests.

We also now terminate the Firefox driver executable immediately rather
than attempting to send a shutdown command. Since the Mozilla
implementation slavishly follows the specification, and there is no server
management API in the spec, nor is there likely to be, no matter how
useful such a thing would be, we will settle for forcing the executable
process to terminate. This is a suboptimal approach, but is required by
the current architecture.