Refactoring detection of spec compliant remote end in .NET
Since the W3C WebDriver working group has decided to remove the mandatory
specificationLevel capability being returned when creating a new session,
we now require a new way to detect whether a remote end uses the
open-source project dialect or the W3C dialect of the wire protocol. We
now do this by sniffing for the 'status' property in responses from the
remote end. OSS dialect responses should have it; W3C dialect responses
should not. Note that this is a potentially flawed algorithm, because
there is no requirement that a remote end cannot add additional fields to
a response body, but it's the best approach at present. If the working
group makes a change that makes it easier to identify when a W3C compliant
version of the protocol is being used, we can switch to that.