Add a check for null results from executing JavaScript
The System.Type.IsInstanceOfType method will always return false when
handed a null value. This change checks explicitly for null to prevent an
exception from being thrown in this case. A check for a value type is also
made to avoid attemping to convert to a value type from a null value.

Fixes #1257