Modifying IE file upload functionality.
Adding additional logging to help diagnose issues where the file
selection dialog is not closed by the driver.

Also adding a new capability, ie.fileUploadDialogTimeout to allow
users to customize the amount of time the driver waits for the file
selection dialog before giving up. It defaults to one second.

Finally, the driver now checks that the file exists before sending
the file name to the file selection dialog. If the file does not
exist, the sendKeys method will throw an exception explicitly stating
that the file does not exist.