Exposing session ID property on RemoteWebDriver in .NET bindings
Users can now cast their driver instance to IHasSessionId to access the
SessionId property, which returns a SessionId object. To use the SessionId
object with external tools, simply call the ToString() method on the
instance. Note that relying on the SessionId for anything is probably a
Bad Thing, because it introduces a leak in the abstraction of a driver.
Please do not use this property; its introduction is the Wrong Approach,
but it's being exposed at the insistence of a selection of users.