Reduces time wasted in when using a native driver
UrlChecker normally adds 1 second to any test that creates and destroys a
DriverService such as the ChromeDriver since the initial poll on both
startup and shutdown nearly always fails and the time between polls
(500 millis) is high.

This switches to using an initially small but exponentially increasing
poll time in order to waste much less time waiting for nothing.

Closes #1334

Signed-off-by: Luke Inman-Semerau <luke.semerau@gmail.com>