Updating textContent attribute test for IE9
The test page is loaded in Quirks mode in IE9, which means that IE doesn't
return a textContent attribute for that page. The alternatives are to
force the test page to load in Standards mode by adding a doctype, or
ignore the test for IE9. This commit chooses the latter approach.