Correcting error handling in .NET finding of elements
This commit removes the `IllegalLocatorException`. This exception was only
thrown when finding by class name where the class name contained
whitespace. It is redundant, and now `InvalidSelectorException` is thrown
in its place. This commit also corrects the element finding tests to
expect the proper `InvalidSelectorException` when finding by tag name,
where the tag name provided is an empty string.