Implementing client-side logging API in .NET
The logging API as implemented in other language bindings and in driver
implementations is suboptimal. The .NET bindings have accordingly resisted
implmenting a client-side API that might be likely to change when proper
API design is done. However, the client-side API in other language
bindings has ossified enough that it's unlikely to change. As such, this
commit implements an API for retrieving logs.