Make IE driver directly return capabilities from new session command
Previously, the driver would return a redirect HTTP response (303), and
the language bindings were expected to follow the redirect to get the
returned capabilities. This has been an obsolete pattern for some time,
and the IE driver is only now conforming to the proper pattern of simply
returning a 200 response with the correct capabilities as part of the
body.