Update IE driver to return NoSuchElement on JavaScript atom errors
Hitting a JavaScript error with the atom is an unrecoverable error. The
most common case of this for IE is when there is a page refresh,
navigation, or similar, and the driver is polling for element presence.
The calling code can't do anything about it, so we might as well just log
and return "no such element" error code. In the common case, this means
that the error will be transitory, and will sort itself out once the DOM
returns to normal after the page transition is completed. Note carefully
that this is an extreme hack, and has the potential to be masking a very
serious problem in the driver.