Fix a test broken by closure update in
f91121d109ac34517763a465463842f5d414e9e8

This test was trivially passing before because errors were
always being caught. The closure update included changes to
the testing framework that detected these types of bugs
and reported them as failures.

I'm not sure what the test was actually trying to exercise,
so I've changed it to what I *think* it was going for.