Rewrite copyFile using Files.copy.
We had some tests observe incomplete copies pretty frequently, and this
seems to fix it. If for some reason this turns out to be noticeably less
efficient than FileChannel, we can rewrite it back to FileChannel and
loop to make sure it copies everything.

(Committed on behalf of a Googler.)