Replace Throwables.propagate with equivalent.
Where safe, replace Throwables.propagate with
equivalent behavior: either "throw e" or
"throw new RuntimeException(e)". This is in
preparation for Guava owners to delete this
method. Note that not all instances of
Throwables.propagate have been removed: those
which could possibly result in the change of
type of the resulting exception.

Signed-off-by: Luke Inman-Semerau <luke.semerau@gmail.com>