Resolve IPv6 addresses in Python remote WebDriver.
The RemoteConnection resolves the hostname in the remote server's
address to an IP in order to avoid the cost of looking up the IP on
every request. This change enables IPv6 resolution; IPv4 is still
preferred, so an IPv6 address is only returned in cases where the host
is not available over IPv4.