Python: Page Object Implementation
This introduces a Page Object implementation in Python.

Pages are defined by PageObject and the WebElements are defined by FindBy and FindAllBy classes.

Upon initialisation a FindBy/FindAllBy class, locators are passed in.

The magic is in the __get__ function, which instead returns the WebElement/list of WebElements instead of the FindBy/FindAllBy. If no element can be found, it will throw a TimeoutException.