Prepare for the maven publish from Buck.
This means that we need to prepare the libraries so that
the dependencies of the maven modules is modelled in the
BUCK files. This involves a certain amount of
jiggery-pokery.

Although this isn't complete, we get a lot closer to being
able to do the libraries under //java/client/src/... with
this change. In order to create the jars correctly, this
will also need a buck version bump.