`./go release` now builds the maven artifacts for release
Well, almost all of them. Not the standalone server.

For some reason, `buck publish` hangs indefinitely, which is
less than ideal. For now we use `buck build` to allow the
process of migration to continue. Interesting, when run out
of the ./go command, `buck publish` works exactly as it
should, so I suspect the problem is somewhere in how we
hooked up the execution of Buck. *sigh*