Optimize ByChained#findElement
Instead of retrieving the entire tree of nested WebElements,
find only the minimum number of elements which allow to give an answer.
In other words, traverse only the left-most path of the aforementioned tree.
This allows for drastically saving on the number of calls to SearchContext#findElements.

Implementation: with a deep-first recursive algorithm,
we scan the whole tree and exit as soon as we find the left-most leaf at the deepest level.

Also, improve the exception message when the bys list is empty.