[atoms] Cleaning up getAttribute dependencies, reducing size from 36K to 7K
The getAttribute atom had a transitive dependency on wgxpath, which was
never being flagged for dead-code removal by the Closure compiler. This
unnecessarily bloats the atom.

As a workaround, move the atom's definition from
webdriver.atoms.element.getAttribute to a new module:
webdriver.atoms.element.attribute (with a single exported "get"
function). Also moving all of the new module's bot.dom dependencies to a
new namespace, bot.dom.core. Having to rearrange the code like this is
a bit unfortunate, but I cannot figure out why the compiler is not pruning
wgxpath from the atom.