[js] Fire the "uncaughtException" event in a new turn of the JS event loop.
As a result of this change, any errors thrown by an event listener will
propagate to the global error handler. Previously, this event was fired with in
the context of a (native) promise callback, causing errors to be silently
suppressed in the promise chain.

This was tested manually. I'm not sure how to test interaction with node's
global error handler using mocha (without crashing the process)

Fixes #2770