[js] Defer wrapping the global mocha functions until first use
This prevents a potential reference error when users run with a --hook
that loads `selenium-webriver/testing`. This creates a cycle with mocha
and the expected globals won't exist when the module runs.

Also adding logic to fail with a more informative message when the
expected globals are not present.

Fixes #2787