add invisibility_of expected condition
I wanted to wait for a progress bar to disappear after a file upload
and I already had the WebElement.  When I found there wasn't an
opposite to visibility_of, I wrote this one since it wasn't a hard
addition and hoped I could contribute it.

I have tested it in my personal code, but if there are any extra test
cases that need to be written for it, I can contribute those too.