py: new FirefoxDriver ctor precedence logic and moz:firefoxOptions support
This change tries to make sense of the keyword arguments accepted by
FirefoxDriver's constructor in the transition phase between the Firefox
add-on driver and the Mozilla supported geckodriver HTTP service.

Prior to this patch, which arguments were consumed and respected varied
greatly on whether the `capabilities["marionette"]` key was set to true.
Following this change, the options passed on to geckodriver will
consistently come out of firefox.Options.

The patch implements precedence logic for `capabilities`,
`firefox_options`, and `firefox_binary`/`firefox_profile`.  They will
populate the underlying data model in that order, which means a binary-
or profile location set in `firefox_options` will override one set in the
`capabilities` dictionary, and that `firefox_binary` and `firefox_profile`
will override both of these.

The goal is to make it predictable what key is chosen, and to make the
underlying data model somewhat easier to understand.  The latter is
particularly important because the extension connection used for the old
Firefox add-on does not accept a binary, profile, or arguments through
capabilities since it is the Python bindings themselves that manage the
lifetime of the browser.

In the case of geckodriver, which is an external HTTP service, it
manages the lifetime and accepts a `moz:firefoxOptions` dictionary that
is not unlike `chromeOptions` as of version 0.11.0.  Versions prior to
geckodriver 0.11.0 accepted `firefox_binary`, `firefox_profile`, and
`firefox_args` as top-level entries on the capabilities JSON Object.
This is no longer supported behaviour by either the Python bindings
or geckodriver.