removing bits of code to allow the original requested new session payload to be forwarded to the node
There's some black magic in grid, where a registered nodes capabilities would get munged with a new session request that was forwarded to the node.

Don't care about removing stuff related to grid v1, that's no longer supported.

This has no affect on grid customizers, beforeSession / beforeCommand still process as they did... but now all of the request is available without it being munged.

This enables users to pass the 'Required Capabilities' in the new session request and it will actually come through.