Allow a Ruby Driver class to be instantiated with an existing session id
Before I get hit for raising this PR, I will try and explain our use case which while
possibly unique, I still believe is valid.

We have an implementation in which we have a bespoke orchestration layer which creates a Selenium
 docker container (including Browsermob) and passes the dynamic url as a connection string to the
test. NB. For various reasons, the orchestration layer is NodeJS while the test is Ruby.

This is an alternative to Selenium Grid which we feel is more stable and maintainable.

Our uses cases are;

1) The orchestration layer provides the session id to the test rather than the current approach of
the orchestration layer querying Selenium API after the test has created the session

2) The orchestration layer can choose Selenium bindings it feels are must suitable during a test run;

e.g. Pass session id to Ruby, do some processing that Ruby is good at, pass session to Javascript binding to
to peform other processing.