Wrapping the session expiration flow with try and catch to prevent from the SessionCleaner thread to exit unexpectedly
We had encounter some situations that an exception can be thrown (mostly runtime exceptions) during the session clean loop and this creates a situation where the SessionCleaner thread stops working and this concludes in a memory leak due to sessions being added and not removed over time.
In general, it is a good idea to put this try and catch clause to support also future code protection.