Wrapping the thread execution flow with try and catch to prevent from the DefaultSession creation to fail unexpectedly
We had encounter some situations that an exception can be thrown (mostly runtime exceptions) during the session creation and this creates a situation where the DefaultSession is not closed as it should and that concludes in a memory leak due to thread pool remain over time. Also the temporary files are not being deleted in that cases.