Allow the BeanToJsonConverter cleanly handle common return types
The way that `toJson` is currently handled encourages people
to take a hard dependency on gson, or attempt to handle
serialising JSON themselves. Modify how `toJson` is handled
to allow us to return other data types, such as `Map` and
anything extending `Collection`.

This opens up the possibility of an infinite loop. Best not
do that in the Real World, eh?