Allow for customisation of TestSlot
The current implementation does not let downstream
consumers of the Grid provide a customised way of
injecting a TestSlot.
This is very useful when one needs to build an On-Demand Grid,
wherein there are no real proxies, but a ghost proxy stands
for the real proxy and based on new session requests,
nodes are spun off.

* Fixed this by defining a method for instantiating a
TestSlot so that downstream RemoteProxy implementations
can choose to provide their own customised TestSlots.

* enhanced the SeleniumProtocol enum to house all
logic related to forming the protocol and the path
from a desired capability into it.

* exposed the matches() method from TestSlot.
Downstream consumers of TestSlot would need access
to alter the logic of matches() method because
sometimes TestSlots would blindly need to return
true for all calls to matches() invocation [ This
is true when one is building a ghost proxy which doesn’t
represent any real node, but stands as a proxy for
something such as a docker daemon from where the
actual sessions would be honoured ]

Without this, end-users would be forced to have their
customized TestSlots reside in the same package
as that of TestSlot (org.openqa.grid.internal)